requirements.txt
amqp
asgiref
async-timeout
billiard
cachetools
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
dj-database-url
Django
django-autoslug
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
flake8
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
kombu
mccabe
oauthlib
pillow
prompt_toolkit
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycodestyle
pyflakes
PyJWT
pyparsing
python-dateutil
redis
requests
requests-oauthlib
rsa
six
sqlparse
stripe
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
pymysql